#6 - PostCommentDto 위치 수정 및 팩토리 메서드 추가
